+++ scripts/hunt-prey/hunted-prey-hook.js
@@ -1,8 +1,37 @@
+import { postToChatConfig } from "../config.js";
 import { HookManager } from "../utils/hook-manager.js";
 import { getFlag, getItem, getItemFromActor } from "../utils/utils.js";
 import { FLURRY_FEATURE_ID, HUNTED_PREY_EFFECT_ID, HUNT_PREY_ACTION_ID, PRECISION_FEATURE_ID, PREY_EFFECT_ID, RANGERS_ANIMAL_COMPANION_FEAT_ID } from "./constants.js";
+import { checkHuntPrey, performHuntPrey } from "./hunt-prey.js";
 
 export function initialiseHuntPrey() {
+
+    // When posting the Hunt Prey chat message, use the Hunt Prey action
+    Hooks.on(
+        "preCreateChatMessage",
+        message => {
+            const item = message.item;
+            if (!item || item.sourceId != HUNT_PREY_ACTION_ID) {
+                return true;
+            }
+
+            const actor = message.actor;
+            if (!actor) {
+                return true;
+            }
+
+            const checkResult = checkHuntPrey(actor);
+            if (!checkResult.valid) {
+                return true;
+            }
+
+            performHuntPrey(actor, null, item, checkResult);
+
+            return game.settings.get("pf2e-ranged-combat", "postActionToChat") == postToChatConfig.full;
+        }
+    );
+
+    // When creating the Hunt Prey effect, also create Hunted Prey effects on the targets
     libWrapper.register(
         "pf2e-ranged-combat",
         "CONFIG.PF2E.Item.documentClasses.effect.prototype._onCreate",
@@ -36,6 +65,7 @@ export function initialiseHuntPrey() {
         "WRAPPER"
     );
 
+    // When deleting the Hunt Prey effect, delete the Hunted Prey effects on the targets
     libWrapper.register(
         "pf2e-ranged-combat",
         "CONFIG.PF2E.Item.documentClasses.effect.prototype._onDelete",
@@ -59,6 +89,7 @@ export function initialiseHuntPrey() {
         "WRAPPER"
     );
 
+    // At the start of the turn, reset the precision feature's attack number back to 1
     Hooks.on(
         "pf2e.startTurn",
         async combatant => {